[DF-6259] Safier timeout values storing.
Former-commit-id: 9f1085954cf9cec695614b54aab4a82b6aea59d9

+++ Sources/Internal/DLC/Downloader/CurlDownloader.h
@@ -170,9 +170,9 @@ private:
         bool IsStarted() {return isStarted;}
         
     private:
-        const int32 timeout;
-        int32 timeLeft;
-        int32 timerStartTime;
+        const int64 timeout;
+        int64 timeLeft;
+        int64 timerStartTime;
         bool isStarted = false;
     };
